fix(review): DSEL 분할과 threat-scan 검증 보정
