+++ mydocs/manual/agent_codex/40_변환과_렌더.md
@@ -2,7 +2,7 @@
 kind: guide
 status: active
 canonical: mydocs/manual/agent_codex/40_변환과_렌더.md
-last_verified: 2026-08-15
+last_verified: 2026-08-16
 generated: tools/gen_agent_codex.py — 수기 수정 금지, 재생성으로 갱신
 ---
 
@@ -112,9 +112,9 @@ rhwp export-svg samples/field-01.hwp -o {tmp}/svg
 ```json
 (비 JSON 출력 — 앞 160자)
 문서 로드 완료: samples/field-01.hwp (3페이지)
-  → <tmp>/svg\field-01_001.svg
-  → <tmp>/svg\field-01_002.svg
-  → <tmp>/svg\field-01_003.svg
+  → <tmp>/svg/field-01_001.svg
+  → <tmp>/svg/field-01_002.svg
+  → <tmp>/svg/field-01_003.svg
 내보내기 완료: 3개 SVG 파일 → <tmp>/sv
 ```
