+++ mydocs/manual/agent_codex/70_자기서술.md
@@ -2,7 +2,7 @@
 kind: guide
 status: active
 canonical: mydocs/manual/agent_codex/70_자기서술.md
-last_verified: 2026-08-15
+last_verified: 2026-08-16
 generated: tools/gen_agent_codex.py — 수기 수정 금지, 재생성으로 갱신
 ---
 
@@ -43,6 +43,19 @@ rhwp export-provenance-map --json
       "note": "앵커 봉투는 경로 에코·해시·연번·머클 루트/경로·판정뿐이다 — 로그와 캡슐은 호출자 산출물이고 문서를 열지 않는다.",
       "origins": {},
       "untrusted": []
+    },
+    "armor": {
+      "note": "safety.nonce·fenceOpen·fenceClose 는 이 호출만의 무작위 격벽 표지(엔진 생성)이고, pageCount·signalCount·clean·scanScopes·safety.note·신호의 종류·주소·근거는 엔진 판정값이다. armoredText 안 격벽 사이 본문… (194자 중 160자)",
+      "origins": {
+        "armoredText": "queries::armor::fence — HwpDocument::extract_page_text_native 로 뽑은 문서 본문을 nonce 격벽으로 감싼 값. 격벽 표지만 엔진 생성이고 격벽 사이 본문은 전부 문서 파생이다",
+        "injectionSignals[].excerpt": "queries::injection_scan::make_excerpt — 주입 신호가 발견된 문서 문맥의 제한 발췌",
+        "injectionSignals[].matched": "queries::injection_scan::scan_text_in — 문서에서 실제 매치된 신호 조각"
+      },
+      "untrusted": [
+        "armoredText",
+        "injectionSignals[].excerpt",
+        "… (3개 중 2개 표시)"
+      ]
     },
     "audit": {
       "note": "감사 봉투는 root(호출자 에코)·개수 회계(total/reproduced/reproducedRate)와 failed[](캡슐 파일 이름·실패 사유·기대/실측 해시)뿐이다 — 캡슐은 문서가 아니라 호출자 산출물이고, 문서 문자열은 재실행 내부에 머문다.",
@@ -940,7 +953,7 @@ rhwp export-agent-manifest --json
         ],
         "summary": "페이지별 텍스트 추출 (TXT 파일 또는 --json stdout)"
       },
-      "… (88개 중 2개 표시)"
+      "… (89개 중 2개 표시)"
     ],
     "exitCodes": {
       "0": "성공",
@@ -2765,6 +2778,19 @@ rhwp export-agent-manifest --json
         "note": "앵커 봉투는 경로 에코·해시·연번·머클 루트/경로·판정뿐이다 — 로그와 캡슐은 호출자 산출물이고 문서를 열지 않는다.",
         "origins": {},
         "untrusted": []
+      },
+      "armor": {
+        "note": "safety.nonce·fenceOpen·fenceClose 는 이 호출만의 무작위 격벽 표지(엔진 생성)이고, pageCount·signalCount·clean·scanScopes·safety.note·신호의 종류·주소·근거는 엔진 판정값이다. armoredText 안 격벽 사이 본문… (194자 중 160자)",
+        "origins": {
+          "armoredText": "queries::armor::fence — HwpDocument::extract_page_text_native 로 뽑은 문서 본문을 nonce 격벽으로 감싼 값. 격벽 표지만 엔진 생성이고 격벽 사이 본문은 전부 문서 파생이다",
+          "injectionSignals[].excerpt": "queries::injection_scan::make_excerpt — 주입 신호가 발견된 문서 문맥의 제한 발췌",
+          "injectionSignals[].matched": "queries::injection_scan::scan_text_in — 문서에서 실제 매치된 신호 조각"
+        },
+        "untrusted": [
+          "armoredText",
+          "injectionSignals[].excerpt",
+          "… (3개 중 2개 표시)"
+        ]
       },
       "audit": {
         "note": "감사 봉투는 root(호출자 에코)·개수 회계(total/reproduced/reproducedRate)와 failed[](캡슐 파일 이름·실패 사유·기대/실측 해시)뿐이다 — 캡슐은 문서가 아니라 호출자 산출물이고, 문서 문자열은 재실행 내부에 머문다.",
